requirements.txt
# Core ML/AI
torch>=2.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0

# PDF Processing
pymupdf4llm>=0.0.10
pymupdf>=1.23.0

# Text Processing
langchain-text-splitters>=0.0.1

# LLM API
google-generativeai>=0.8.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0

# OCR (using Gemini vision)
Pillow

# Optional: Web UI
# gradio>=4.0.0
